+++ src/env.js
@@ -11,8 +11,12 @@ export const env = createEnv({
 			process.env.NODE_ENV === "production"
 				? z.string()
 				: z.string().optional(),
+
 		AUTH_GOOGLE_CLIENT_ID: z.string(),
 		AUTH_GOOGLE_CLIENT_SECRET: z.string(),
+
+		GOOGLE_API_KEY: z.string(),
+
 		DATABASE_URL: z.string().url(),
 		NODE_ENV: z
 			.enum(["development", "test", "production"])
@@ -37,6 +41,8 @@ export const env = createEnv({
 
 		AUTH_GOOGLE_CLIENT_ID: process.env.AUTH_GOOGLE_CLIENT_ID,
 		AUTH_GOOGLE_CLIENT_SECRET: process.env.AUTH_GOOGLE_CLIENT_SECRET,
+
+		GOOGLE_API_KEY: process.env.GOOGLE_API_KEY,
 
 		DATABASE_URL: process.env.DATABASE_URL,
 		NODE_ENV: process.env.NODE_ENV,