+++ src/server/api/root.ts
@@ -1,4 +1,4 @@
-import { postRouter } from "~/server/api/routers/post";
+import { magicRouter } from "~/server/api/routers/magic";
 import { createCallerFactory, createTRPCRouter } from "~/server/api/trpc";
 
 /**
@@ -7,7 +7,7 @@ import { createCallerFactory, createTRPCRouter } from "~/server/api/trpc";
  * All routers added in /api/routers should be manually added here.
  */
 export const appRouter = createTRPCRouter({
-	post: postRouter,
+	magic: magicRouter,
 });
 
 // export type definition of API
@@ -17,7 +17,5 @@ export type AppRouter = typeof appRouter;
  * Create a server-side caller for the tRPC API.
  * @example
  * const trpc = createCaller(createContext);
- * const res = await trpc.post.all();
- *       ^? Post[]
  */
 export const createCaller = createCallerFactory(appRouter);