+++ src/server/api/routers/hello.ts
@@ -1,0 +1,7 @@
+import { createTRPCRouter, publicProcedure } from "~/server/api/trpc";
+
+export const helloRouter = createTRPCRouter({
+	hello: publicProcedure.query(() => {
+		return "Hello, world!";
+	}),
+});